+++ sismic/testing/__init__.py
@@ -1,3 +1,3 @@
-from .tester import teststory_from_trace
+from .tester import teststory_from_trace, ExecutionWatcher
 
 __all__ = ['teststory_from_trace']