+++ tests/test_evaluator.py
@@ -67,3 +67,10 @@ class PythonEvaluatorTests(unittest.TestCase):
     def test_send(self):
         self.evaluator._execute_code('send("hello")')
         self.interpreter.queue.assert_called_with(InternalEvent('hello'))
+
+    def test_add_variable_in_context(self):
+        self.evaluator._execute_code('a = 1\nassert a == 1')
+        self.assertTrue(self.evaluator._evaluate_code('a == 1'))
+
+    def test_access_outer_scope(self):
+        self.evaluator._execute_code('a = 1\nd = [x for x in range(10) if x!=a]')